+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # Changelog
 
+## v0.0.10
+
+- Change configuration file format. Now, the configuration file is JS instead of JSON. This provides better discoverability and type checking for TypeScript users.
+
+The following `.protoscript.json`:
+
+```json
+{
+  "root": "src",
+};
+```
+
+Would be renamed to `proto.config.mjs` and changed to the following:
+
+```json
+/** @type {import('protoscript').Config} */
+export default {
+  root: "src",
+};
+```
+
 ## v0.0.9
 
 - Remove `process.stdin.fd` usage to see if it resolves intermittent `Error: EAGAIN: resource temporarily unavailable, read`. See [#191](https://github.com/tatethurston/TwirpScript/issues/191) for more context.